Fix for document_translations table (#228)

+++ db/migrate/20180103185615_drop_documents.rb
@@ -1,6 +1,6 @@
 class DropDocuments < ActiveRecord::Migration
   def change
     drop_table :documents
-    drop_table :document_translations
+    drop_table :document_translations if table_exists?(:document_translations)
   end
 end
